Picking radius works. Can pick points and lines now!
Added a new demo, PointsAndLines, to show this in action. Also
modified the ThreeObjects demo to use a positive picking radius (just
for testing purposes).

As a side note, the "ignore back faces" doesn't work when the pick
radius is not zero. I marked this as a bug, even though I don't know
if can fix it (the problem is that the PolytopeIntersector does not
provide the intersection normals).